requirements.txt
transformers
accelerate
sentence-transformers
faiss-cpu
PyYAML
fastapi==0.111.0
uvicorn[standard]==0.30.1fastapi==0.111.0
uvicorn[standard]==0.30.1
unstructured
python-magic
lxml
chromadb
sentence_transformers
pypdf
langchain_text_splitters